requirements.txt
django
djangorestframework
docker
markdown
django-filter
Pillow
django-rest-knox
psycopg2-binary